third_party/blink/web_tests/fast/backgrounds/background-shorthand-after-set-backgroundSize-expected.txt

Tests a flag to make background shorthand property not override background-size when it's already set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.style.background is "url(\"dummy://test.png\") center center / cover no-repeat border-box border-box red"
PASS e.style.backgroundSize is "cover"


PASS successfullyParsed is true

TEST COMPLETE
